requirements.txt
discord
discord_webhook
aiohttp
pyyaml
Pillow
jishaku
colorama